chore: drop Snap packaging
We do not have a lot of expertise dealing with Snap and as it is
currently breaking our CI, it might be good to drop support for it.

This does not mean that it won't come back in the future, but as it
stands, it was effectively unmaintained and was only used for nightly
builds anyway.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>